+++ domains/genericPoker.h
@@ -136,7 +136,7 @@ class GenericPokerState : public State {
   // Destructor
   ~GenericPokerState() override = default;
 
-  int countAvailableActionsFor(Player player) const override;
+  unsigned long countAvailableActionsFor(Player player) const override;
 
   // GetActions returns possible actions for a player in the state.
   vector<shared_ptr<Action>> getAvailableActionsFor(Player player) const override;
